requirements.txt
pypdf==3.17.1
langchain==0.3.18
langchain-chroma==0.2.1
langchain-community==0.3.17
langchain-core==0.3.34
langchain-ollama==0.2.3
chromadb==0.6.3
pytest==8.1.1
boto3==1.34.69
streamlit==1.32.2
python-dotenv==1.0.1
typing-extensions>=4.5.0
black==24.3.0
pylint==3.1.0
isort==5.13.2
mypy==1.9.0

